+++ decision/decision.go
@@ -8,70 +8,7 @@ import (
 
 	"github.com/flagship-io/flagship-common/utils"
 	"github.com/flagship-io/flagship-proto/decision_response"
-	"github.com/golang/protobuf/ptypes/wrappers"
-	"google.golang.org/protobuf/types/known/structpb"
 )
-
-// getPreviousABVGIds returns previously assigned AB test campaigns for visitor
-func getPreviousABVGIds(variationGroups []*VariationsGroup, existingVar map[string]*VisitorVGCacheItem) []string {
-	previousVisVGsAB := []string{}
-	for _, vg := range variationGroups {
-		if vg.CampaignType != "ab" {
-			continue
-		}
-		existingVariations, ok := existingVar[vg.ID]
-		isActivated := ok && existingVariations.Activated
-		if isActivated && !utils.IsInStringArray(vg.ID, previousVisVGsAB) {
-			previousVisVGsAB = append(previousVisVGsAB, vg.ID)
-			break
-		}
-	}
-	return previousVisVGsAB
-}
-
-// buildCampaignResponse creates a decision campaign response, filling out empty flag keys for each variation if needed
-func buildCampaignResponse(vg *VariationsGroup, variation *Variation, shouldFillKeys bool) *decision_response.Campaign {
-	campaignResponse := decision_response.Campaign{
-		Id: &wrappers.StringValue{
-			Value: vg.CampaignID,
-		},
-		VariationGroupId: &wrappers.StringValue{
-			Value: vg.ID,
-		},
-	}
-
-	if shouldFillKeys {
-		if variation.Modifications == nil {
-			variation.Modifications = &decision_response.Modifications{}
-		}
-		if variation.Modifications.Value == nil {
-			variation.Modifications.Value = &structpb.Struct{}
-		}
-		if variation.Modifications.Value.Fields == nil {
-			variation.Modifications.Value.Fields = map[string]*structpb.Value{}
-		}
-		for _, v := range vg.Variations {
-			if v.Modifications != nil && v.Modifications.Value != nil && v.Modifications.Value.Fields != nil {
-				for key := range v.Modifications.Value.Fields {
-					if _, ok := variation.Modifications.Value.Fields[key]; !ok {
-						variation.Modifications.Value.Fields[key] = &structpb.Value{Kind: &structpb.Value_NullValue{}}
-					}
-				}
-			}
-		}
-	}
-
-	protoModif := &decision_response.Variation{
-		Id: &wrappers.StringValue{
-			Value: variation.ID,
-		},
-		Modifications: variation.Modifications,
-		Reference:     variation.Reference,
-	}
-
-	campaignResponse.Variation = protoModif
-	return &campaignResponse
-}
 
 // GetDecision return a decision response from visitor & environment infos
 func GetDecision(
