+++ decision/models.go
@@ -1,15 +1,11 @@
 package decision
 
 import (
-	"fmt"
-	"log"
-	"sort"
 	"time"
 
 	"github.com/flagship-io/flagship-common/utils"
 	"github.com/flagship-io/flagship-proto/decision_response"
 	"github.com/flagship-io/flagship-proto/targeting"
-	protoStruct "github.com/golang/protobuf/ptypes/struct"
 	"google.golang.org/protobuf/types/known/structpb"
 )
 
@@ -135,64 +131,4 @@ func (s byCreatedAtVG) Swap(i, j int) {
 }
 func (s byCreatedAtVG) Less(i, j int) bool {
 	return s[i].CreatedAt.Before(s[j].CreatedAt)
-}
-
-// GetCampaignsArray returns the first campaign that matches the campaign ID
-func GetCampaignsArray(campaigns map[string]*CampaignInfo) []*CampaignInfo {
-	// Use ID array to sort variation groups by ID to force order of iteration
-	cArray := []*CampaignInfo{}
-	cIDs := map[string]bool{}
-	for _, c := range campaigns {
-		if _, ok := cIDs[c.ID]; !ok {
-			cArray = append(cArray, c)
-		}
-		cIDs[c.ID] = true
-
-	}
-	sort.Sort(byCreatedAtCampaigns(cArray))
-	return cArray
-}
-
-// GetVariationGroup returns the first variationGroup that matches the visitorId and context
-func GetVariationGroup(variationGroups map[string]*VariationsGroup, visitorID string, context map[string]*protoStruct.Value) *VariationsGroup {
-	// Use ID array to sort variation groups by ID to force order of iteration
-	vgArray := []*VariationsGroup{}
-	for _, vg := range variationGroups {
-		vgArray = append(vgArray, vg)
-	}
-	sort.Sort(byCreatedAtVG(vgArray))
-
-	for _, variationGroup := range vgArray {
-		match, err := TargetingMatch(variationGroup, visitorID, context)
-		if err != nil {
-			log.Println(fmt.Sprintf("Targeting match error variationGroupId %s, user %s: %s", variationGroup.ID, visitorID, err))
-		}
-		if match {
-			return variationGroup
-		}
-	}
-	return nil
-}
-
-// GetCampaignsVG returns the variation groups that target visitor
-func GetCampaignsVG(campaigns []*CampaignInfo, visitorID string, context map[string]*structpb.Value) []*VariationsGroup {
-	campaignVG := []*VariationsGroup{}
-	existingCampaignVG := make(map[string]bool)
-	for _, campaign := range campaigns {
-		_, ok := existingCampaignVG[campaign.ID]
-		if ok {
-			// Campaign already handled (maybe because of custom ID)
-			continue
-		}
-
-		vg := GetVariationGroup(campaign.VariationsGroups, visitorID, context)
-
-		if vg == nil {
-			continue
-		}
-		vg.CampaignType = campaign.Type
-		existingCampaignVG[campaign.ID] = true
-		campaignVG = append(campaignVG, vg)
-	}
-	return campaignVG
 }
